Changes from NB project
Set products as abstract for categories, which will each have its own
implementation of add/remove with some special considions such a perish
slightly(groceries), or be stolen(hardware), or normal(plumbing).
Inventory can be a List holding class with some sort of set inventory
limit on the number of items. Display renders the number of items and
buttons to remove/add different items